@@ -1,39 +1,18 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.Linq;
-using System.Text;
-using System.Text.RegularExpressions;
-using System.Threading.Tasks;
+﻿using System.Text.RegularExpressions;
 
-namespace UtilityLibrary.LanguageExtensions
+namespace UtilityLibrary.LanguageExtensions;
+
+/// <summary>
+/// Common string extensions 
+/// </summary>
+public static class StringExtensions
 {
+
     /// <summary>
-    /// Common string extensions 
+    /// Use to split on upper cased characters and separate with a single space.
     /// </summary>
-    public static class StringExtensions
-    {
+    public static string SplitCamelCase(this string sender) =>
+        string.Join(" ", Regex.Matches(sender, @"([A-Z][a-z]+)")
+            .Select(m => m.Value));
 
-        /// <summary>
-        /// Use to split on upper cased characters and separate with a single space.
-        /// </summary>
-        public static string SplitCamelCase(this string sender) =>
-            string.Join(" ", Regex.Matches(sender, @"([A-Z][a-z]+)")
-                .Select(m => m.Value));
-
-        /// <summary>
-        /// Determine if string is empty
-        /// </summary>
-        /// <param name="sender">String to test if null or whitespace</param>
-        /// <returns>true if empty or false if not empty</returns>
-        [DebuggerStepThrough]
-        public static bool IsNullOrWhiteSpace(this string sender) => string.IsNullOrWhiteSpace(sender);
-        /// <summary>
-        /// Determine if a string can be represented as a numeric.
-        /// </summary>
-        /// <param name="text">value to determine if can be converted to a string</param>
-        /// <returns></returns>
-        [DebuggerStepThrough]
-        public static bool IsNumeric(this string text) => double.TryParse(text, out _);
-    }
 }
